When sorting on priority, sort also on projects and contexts
After pressing <leader>s, todo.vim used to sort all lines
alphabetically. This caused tasks related to different projects to be
placed apart from each other - the whole list was then difficult to
follow.

This patch fixes this by taking into account the fact that :sort in vim
is usually stable (see http://en.wikipedia.org/wiki/Sorting_algorithm#Stability),
so that when we sort first on contexts, then on projects and then on
priority, everything is placed in the order we want.